BATTLE. Fix destroy alien.
Super hacky fix for destroy alien. Using gsub on emojis proved to be quite difficult and requires a regex. I was unable to find the correct value for our emoji, so I am just replacing the string instead of subbing out the alien emoji. This can be fixed by finding the correct regex for out alien emoji