GAME LEVEL: Add a level module, import into gameLoop.
Add a level module that will keep track of the level. As of now, it just prints the player ship, in the future it should read and write to a file to get the <level_state>? States aren't allowed in Erlang, but we can do something with this and pass it as a message rather than use it as a state.